# docs/swagger.yaml
basePath: /
definitions:
...
  model.Product:
    properties:
      amount:
        type: integer
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      imageID:
        type: string
      order_id:
        type: integer
      price:
        type: number
      purchased_amount:
        type: integer
      related_products:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      reviews:
        items:
          $ref: '#/definitions/model.Review'
        type: array
      tag:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      views:
        type: integer
    required:
    - amount
    - category
    - price
    - title
    type: object
  model.Review:
    properties:
      category:
        type: string
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      text:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    required:
    - category
    - text
    type: object
...
  v1.cartInput:
    properties:
      amount:
        type: integer
    required:
    - amount
    type: object
  v1.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  v1.getProductsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Product'
        type: array
    type: object
...
  v1.statusResponse:
    properties:
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Simple market API
  title: Market API
  version: "1.0"
paths:
  /api/v1/cart:
    delete:
      operationId: delete-products-from-cart
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete products from cart
      tags:
      - cart
    get:
      operationId: get-products-from-cart
      parameters:
      - description: sort by
        enum:
        - created_at
        in: query
        name: sort_by
        type: string
      - description: sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: limit
        enum:
        - 10
        - 25
        - 50
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.getProductsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get products from cart
      tags:
      - cart
  /api/v1/cart/{productId}:
    delete:
      operationId: delete-product-from-cart
      parameters:
      - description: ID of product to delete
        in: path
        name: productId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.getProductsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete product from cart
      tags:
      - cart
    post:
      consumes:
      - application/json
      operationId: add-product-to-cart
      parameters:
      - description: ID of product to add to cart
        in: path
        name: productId
        required: true
        type: integer
      - description: Amount of products
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.cartInput'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.getProductsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add product to cart
      tags:
      - cart
    put:
      consumes:
      - application/json
      operationId: update-product-amount-from-cart
      parameters:
      - description: ID of product to update
        in: path
        name: productId
        required: true
        type: integer
      - description: Amount of products
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.cartInput'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.getProductsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update product amount from cart
      tags:
      - cart
...
  /api/v1/product/{productId}:
    delete:
      operationId: delete-product
      parameters:
      - description: ID of product to delete
        in: path
        name: productId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete product from the market
      tags:
      - products
    get:
      operationId: get-product-by-id
      parameters:
      - description: ID of product to get
        in: path
        name: productId
        required: true
        type: integer
      - description: sort by
        enum:
        - created_at
        in: query
        name: sort_by
        type: string
      - description: sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: limit
        enum:
        - 10
        - 25
        - 50
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      summary: Get product by id from the market
      tags:
      - products
    put:
      consumes:
      - multipart/form-data
      operationId: update-product
      parameters:
      - description: ID of product to update
        in: path
        name: productId
        type: integer
      - description: Image to Upload
        in: formData
        name: file
        type: file
      - description: Title of product
        in: formData
        name: title
        type: string
      - description: Price of product
        in: formData
        name: price
        type: number
      - description: Tag of product
        in: formData
        name: tag
        type: string
      - description: Category of product
        in: formData
        name: category
        type: string
      - description: Description of product
        in: formData
        name: description
        type: string
      - description: Amount of products
        in: formData
        name: amount
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing product from the market
      tags:
      - products
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"